Refactor test for login
